# lib/db/src/schema/alerts.ts
import { pgTable, text, timestamp, doublePrecision } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const alertsTable = pgTable("alerts", {
  id: text("id").primaryKey(),
  type: text("type", { enum: ["sos", "fall", "tamper", "manual"] }).notNull(),
  status: text("status", { enum: ["active", "resolved"] }).notNull().default("active"),
  latitude: doublePrecision("latitude").notNull(),
  longitude: doublePrecision("longitude").notNull(),
  message: text("message"),
  resolvedAt: timestamp("resolved_at", { withTimezone: true }),
  timestamp: timestamp("timestamp", { withTimezone: true }).notNull().defaultNow(),
});

export const insertAlertSchema = createInsertSchema(alertsTable);
export type InsertAlert = z.infer<typeof insertAlertSchema>;
export type Alert = typeof alertsTable.$inferSelect;